refactor(mobile): split components
